// src/jobs/types/enums.ts
export enum JobType {
    saleListener = 'saleListener',
    createdContractListener = 'createdContractListener',
    createdTokenListener = 'createdTokenListener',
    burnedTokenListener = 'burnedTokenListener',
    transferTokenListener = 'transferTokenListener',
    launchpadSaleListener = 'launchpadSaleListener',
}